Fixed a likely typo
Sorry if I'm wrong here, since I'm not very experienced with Elixir, but it looks like Socket.Datagram.send! is calling a function in Socket.Stream.Protocol rather than Socket.Datagram.Protocol. Not a major issue, since Socket.Datagram.send (without the bang) is equivalent, but it's a little quirk that forced me to look up the implementation.