Fix `recv` when length is zero in `Socket.Web`
`:gen_tcp.recv/3` and `:ssl.recv/3` function, when receiving a length 0,
will both try to fetch all available bytes, which is not suitable for
websocket case.

Eg. when the server sends a ping with no payload, we should not try to
fetch bytes from the server. Instead, we should respond with `pong`
as soon as possible. (of course, the responding thing is the job
of `elixir-socket` user)

Also, when reading the codes, I found a potential bug for calling
`forge` function. Before, if `options[:mask] = nil`, it will never
be passed to `forge` function with `options[:mask] || mask` clause.